Added new merge algorithm InfoFlowMap, which first performs InfoFlow, then performs InfoMap.
Applying InfoFlowMap to existing graphs suggests no InfoMap merging can be performed after InfoFlow, i.e., after InfoFlow, not one pair of modules can be merged to reduce code length.

To faciliate logging to the same log file with InfoFlow and InfoMap, an append option is added to LogFile. This option is not accessible from the config file, so that from the users' perspective, the log file does not have an append option.